+++ AoC/ITestableSolver.cs
@@ -2,7 +2,7 @@
 // 
 //  AocAutomaton
 // 
-//  Copyright (c) 2023 Cyrille DUPUYDAUBY
+//  Copyright (c) 2024 Cyrille DUPUYDAUBY
 // ---
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -22,45 +22,11 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using System;
-
 namespace AoC;
 
-public class TestData
+public interface ITestableSolver : ISolver
 {
-    public TestData(string data, string init = null)
-    {
-        Data = data;
-        Init = init;
-    }
-    
-    public string Data { get; }
-    
-    public string Init { get; }
-    
-    public object[] Answers { get; } = new object[2];
+    bool IsTest { get; set; }
 
-    public bool[] VisualConfirm { get; } = new bool[2];
-    
-    public Action[] SetupActions { get; } = new Action[2];
-
-    public bool CanTest(int id) => Answers[id - 1] != null || VisualConfirm[id-1];
-
-    public TestData Answer1(object answer)
-    {
-        Answers[0] = answer;
-        return this;
-    }
-    
-    public TestData Answer2(object answer)
-    {
-        Answers[1] = answer;
-        return this;
-    }
-
-    public TestData SetVisualConfirm(int question)
-    {
-        VisualConfirm[question-1] = true;
-        return this;
-    }
+    bool UnitTest();
 }