+++ AoC/SolverWithParser.cs
@@ -30,7 +30,6 @@ namespace AoC
 
         private string Data
         {
-            get => _data;
             set
             {
                 if (value == _data) return;
@@ -46,6 +45,16 @@ namespace AoC
         
         public abstract void SetupRun(Automaton automatonBase);
 
+        /// <summary>
+        ///     Parse the exercise data.
+        /// </summary>
+        /// <param name="data">data to be parsed as a single string</param>
+        /// <remarks>
+        ///     Split the data as an array of lines and call <see cref="SolverWithLineParser.ParseLine" /> for each them.
+        ///     Override this method if you prefer to parse data as a whole (you still need to provide an empty ParseLine method
+        ///     as it is declared abstract).
+        /// </remarks>
+        protected abstract void Parse(string data);
         public object GetAnswer1(string data)
         {
             Data = data;
@@ -61,16 +70,5 @@ namespace AoC
         public abstract object GetAnswer1();
         
         public abstract object GetAnswer2();
-
-        /// <summary>
-        ///     Parse the exercise data.
-        /// </summary>
-        /// <param name="data">data to be parsed as a single string</param>
-        /// <remarks>
-        ///     Split the data as an array of lines and call <see cref="SolverWithLineParser.ParseLine" /> for each them.
-        ///     Override this method if you prefer to parse data as a whole (you still need to provide an empty ParseLine method
-        ///     as it is declared abstract).
-        /// </remarks>
-        protected abstract void Parse(string data);
     }
 }